Added StartDetector and EndRaceDetector classes. Also added a global variable.
The global variable isStarted is so that multiple classes can access is without
requiring inheritance.

StartDetector class is a child of Detector. Only thing in it is the handle
method, which Detector.process transfers control to if detection is positive.
If the start of the race is detected, set global isStarted to True.

EndRaceDetector class is used for detecting the end of a race for phase 0
splitting purposes. Tests to see when the horizontal line disappears from the
frame. Detection is bypassed if the race has not started, i.e. whenever the
current frame is not before the race start. Once the race begins, detection
starts. When it is below some threshold, now 600, transfer control to handle
method. The handle method will reset global isStarted to False.