Small changes to detection.py and changed up the structure of phase_0.py
In detection.py I made it so that it imports the database module. In
EndRaceDetector.handle there is also a call to database.put_race which
will pass off race start and duration, given a session id.

In phase_0.py, the main block has been made a method so that server.py can
start phase_0. A main block is still in existence for debugging purposes.