Changed EndRaceDetector detection algorithm.
Instead of thresholding the whole frame in color, like before, the script converts the frame to
grayscale and sets the verticle black player screen separator as the ROI, and then thresholds that.
We then get the number of pixels that are equal to 0, and execution is the same after that. This
dramatically improves the speed of phase 0.