Added PlayerNum support to detection.py StartRaceDetector
Changed the flow of Detector.detect() and StartRaceDetector.handle().
In Detector.detect(), we know iterate through the different player regions
and process on each individual region. We get the regions from BoxExtractor (soon to come).
We then record the number of starts detected, based on each region fed to Detector.process().
The number of starts directly corresponds to the number of players.

In StartRaceDetector.handle(), we just added a numStarts variable that keeps
track of the number of hits in 1 frame.