Altered finish masks and added new position masks
Finish masks have been converted from RGB images to binary images. The method
of conversion was binary thresholding.

Position masks were generated in the same way. There are currently more masks
then will be eventually necessary, but this is ok for now. We can delete a
mask or two when we have a better idea of which ones perform better.