Altered the default_shape variable to work with a list of default shapes
Within phase_0 and phase_1, we now pass default_shape as a list. This was done
to work with varying source size for mask acquisition.

Now, the Detector base class checks to see if the length of default_shape is 1 or
the number of masks passed to the detector. For any detector that has a defined
process method, support for the default_shape list has been added, such that each
mask has it's own default_shape for scaling.

Additionally, there is a new zoomed out start mask.