Fixed event creation and processing for phase_1 detectors
Added an is_started field inside the players object for the shortcut detector.
Also added an events list for creating events and appending them.

In processing.py, we now have correct handling for event creation.

In both genric.py and processing.py, a fix has been implemented for the process()
method (for the generic Detector.process() and all custom process() methods).
The problem was that zipping the masks with a default_shape of length 1 caused
only 1 mask to be iterated through. Now there is a quick check of the length
of self.default_shape. If it's 1, go to 1 for loop. If not, go to the other.

Changes in phase_1.py were simply to account for changes made in other files.
A print of the events list was also added for debugging.