Large changes/fixes for various detectors. Implemented blurring for all detectors
In config, we now hardcode the frame rate to float(30).

In utility, the RingBuffer.exists() and RingBuffer.find_unique() have been
altered to work with Chars.

In parallel.py, added debugging functionality for fast testing. We can now
press the spacebar key in order to toggle all active detectors.

In processing, all instances of blurring have been removed, except for in Shortcut.
This is because Detector now does it by default. A few kinks in Items hae been sorted
out as well.

In splitting, Chars has been fixed. In its previous implementation, it just didn't
work. Now it locks in the characters correctly and provides a correct number
for number of players.