Added Map detector in detection/splitting.py. Also have new map masks
Map detection is now using a binary threshold approach on the map select screen.
The map is locked in whenever the BlackFrame detector sees a black frame.

****************************BIG TODO********************************
Currently in phase_0, we lock in player regions at the start of a race.
However, map detector needs the extracted box from a single box frame.
By that, I mean that map detector must see the output from BoxExtractor, since
it must see the entirety of the map select frame (with the extracted box from
BoxExtractor). This works for the first race in a session, but not subsequent races.