Added Fall and Reverse detectors with masks. Fixed Shortcut.
Fall detectors detects when a player has fallen off the map. Since its mask is so
similar to any generic lakitu mask, the threshold has been moved to 0.02.

Reverse detects when a player is going in reverse

Shortcut now has a smaller threshod to kill false positives for a short while